Refactor - Separated out ResultsPanel
